Rename the components, fix ripple effect and created a password text field
