+++ nbproject/project.properties
@@ -18,8 +18,12 @@ build.test.results.dir=${build.dir}/test/results
 #debug.transport=dt_socket
 debug.classpath=\
     ${run.classpath}
+debug.modulepath=\
+    ${run.modulepath}
 debug.test.classpath=\
     ${run.test.classpath}
+debug.test.modulepath=\
+    ${run.test.modulepath}
 # Files in build.classes.dir which should be excluded from distribution jar
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
@@ -34,6 +38,8 @@ javac.classpath=
 javac.compilerargs=
 javac.deprecation=false
 javac.external.vm=true
+javac.modulepath=
+javac.processormodulepath=
 javac.processorpath=\
     ${javac.classpath}
 javac.source=1.8
@@ -41,6 +47,8 @@ javac.target=1.8
 javac.test.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
+javac.test.modulepath=\
+    ${javac.modulepath}
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
@@ -66,9 +74,13 @@ run.classpath=\
 # You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
 # To set system properties for unit tests define test-sys-prop.name=value:
 run.jvmargs=
+run.modulepath=\
+    ${javac.modulepath}
 run.test.classpath=\
     ${javac.test.classpath}:\
     ${build.test.classes.dir}
+run.test.modulepath=\
+    ${javac.test.modulepath}
 source.encoding=UTF-8
 src.dir=src
 test.src.dir=test
