+++ src/com/hq/swingmaterialdesign/JFrame_Main.form
@@ -41,7 +41,7 @@
   <SubComponents>
     <Container class="javax.swing.JPanel" name="jPanel_MainPanel">
       <AuxValues>
-        <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialPanel()"/>
+        <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MPanel()"/>
       </AuxValues>
 
       <Layout>
@@ -148,7 +148,7 @@
                 <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jButton_CloseActionPerformed"/>
               </Events>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialButton()"/>
+                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MButton()"/>
               </AuxValues>
             </Component>
             <Component class="javax.swing.JButton" name="jButton_Maximize">
@@ -172,7 +172,7 @@
                 <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jButton_MaximizeActionPerformed"/>
               </Events>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialButton()"/>
+                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MButton()"/>
               </AuxValues>
             </Component>
             <Component class="javax.swing.JButton" name="jButton_Minimize">
@@ -196,7 +196,7 @@
                 <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jButton_MinimizeActionPerformed"/>
               </Events>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialButton()"/>
+                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MButton()"/>
               </AuxValues>
             </Component>
             <Component class="javax.swing.JLabel" name="jLabel1">
@@ -212,7 +212,7 @@
         </Container>
         <Container class="javax.swing.JPanel" name="jPanel_Main">
           <AuxValues>
-            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialPanel()"/>
+            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MPanel()"/>
           </AuxValues>
 
           <Layout>
@@ -270,7 +270,7 @@
                 </Property>
               </Properties>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialComboBox&lt;&gt;()"/>
+                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MComboBox&lt;&gt;()"/>
                 <AuxValue name="JavaCodeGenerator_TypeParameters" type="java.lang.String" value="&lt;String&gt;"/>
               </AuxValues>
             </Component>
@@ -279,7 +279,7 @@
                 <Property name="text" type="java.lang.String" value="Swing Material Design"/>
               </Properties>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialTextField()"/>
+                <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MTextField()"/>
               </AuxValues>
             </Component>
           </SubComponents>
@@ -303,7 +303,7 @@
             <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jButton_MainActionPerformed"/>
           </Events>
           <AuxValues>
-            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialButton()"/>
+            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MButton()"/>
           </AuxValues>
         </Component>
         <Component class="javax.swing.JButton" name="jButton_Main1">
@@ -325,7 +325,7 @@
             <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jButton_Main1ActionPerformed"/>
           </Events>
           <AuxValues>
-            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new MaterialButton()"/>
+            <AuxValue name="JavaCodeGenerator_CreateCodeCustom" type="java.lang.String" value="new com.hq.swingmaterialdesign.materialdesign.MButton()"/>
           </AuxValues>
         </Component>
         <Container class="javax.swing.JToolBar" name="jToolBar_Status">
