+++ src/com/hq/swingmaterialdesign/materialdesign/RippleEffect.java
@@ -38,7 +38,7 @@ public class RippleEffect {
             int rippleRadius = (int) ripple.getRippleRadius();
             Color fg = g2.getColor();
             g2.setColor(new Color(fg.getRed() / 255f, fg.getGreen() / 255f, fg.getBlue() / 255f, rippleOpacity));
-            g2.fillOval(rippleCenter.x - rippleRadius, rippleCenter.y - rippleRadius, 2 * rippleRadius, 2 * rippleRadius);
+            g2.fillOval(rippleCenter.x - rippleRadius, rippleCenter.y - rippleRadius, rippleRadius, rippleRadius);
         }
     }
